requirements.txt
arrow
axe-playwright-python
ephemeral-port-reserve
fastapi
langchain
langchain-community
langchain-openai
langchain-huggingface
markdown
more-itertools
ollama
opensearch-py
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
playwright
psycopg
psycopg-binary
pytailwindcss
pytest
pytest-playwright
pytest-xdist
python-dotenv
redis
sentence-transformers
sqlmodel